some fixes

+++ src/client/app/components/carousel/photo-slider.component.ts
@@ -14,7 +14,7 @@ export class PhotoSliderComponent implements OnInit {
 
   filesToUpload: FileList;
   userId: number;
-  photos: Photo[] = [];
+  photos: Photo[] = [ {userId: 0, _id: '', base64: '', avatar: false, name: ''} ];
 
   constructor(private photosService: PhotosService) {}
 
@@ -35,12 +35,14 @@ export class PhotoSliderComponent implements OnInit {
   // }
 
   deletePhoto(photoId: string) {
-    this.photosService.deletePhoto(photoId).subscribe(() => {
-      this.photosService.getPhotos(this.userId)
-        .subscribe(items => {
-          this.photos = items;
-        });
-    });
+    if (confirm('Delete photo?')){
+      this.photosService.deletePhoto(photoId).subscribe(() => {
+        this.photosService.getPhotos(this.userId)
+          .subscribe(items => {
+            this.photos = items;
+          });
+      });
+    } else { return; }
   }
 
   fileChangeEvent(fileInput: any) {
